Review feedback: java DSL added, and other minor changes.
