+++ akka-persistence-typed/src/main/scala/akka/persistence/typed/internal/BehaviorSetup.scala
@@ -15,6 +15,7 @@ import akka.annotation.InternalApi
 import akka.persistence._
 import akka.persistence.typed.EventAdapter
 import akka.persistence.typed.PersistenceId
+import akka.persistence.typed.RetentionCriteria
 import akka.persistence.typed.scaladsl.EventSourcedBehavior
 import akka.util.ConstantFun
 import akka.util.OptionVal
@@ -124,59 +125,4 @@ private[akka] object MDC {
   def create(persistenceId: PersistenceId, phaseName: String): Map[String, Any] = {
     Map("persistenceId" -> persistenceId.id, "phase" -> phaseName)
   }
-}
-
-/**
-  * INTERNAL API
-  * Setup snapshot and event delete/retention behavior. Retention bridges snapshot
-  * and journal behavior. This defines the retention criteria.
-  *
-  * @param snapshotEveryNEvents Snapshots are used to reduce playback/recovery times.
-  *                             This defines when a new snapshot is persisted.
-  *
-  * @param keepNSnapshots      After a snapshot is successfully completed,
-  *                             - if 2: retain last maximum 2 *`snapshot-size` events
-  *                             and 3 snapshots (2 old + latest snapshot)
-  *                             - if 0: all events with equal or lower sequence number
-  *                             will not be retained.
-  *
-  * @param deleteEventsOnSnapshot Opt-in ability to delete older events on successful
-  *                               save of snapshot. Defaults to disabled.
-  */
-final case class RetentionCriteria(snapshotEveryNEvents: Long,
-                                   keepNSnapshots: Long,
-                                   deleteEventsOnSnapshot: Boolean) {
-  /**
-    * Delete Messages:
-    *   {{{ toSequenceNr - keepNSnapshots * snapshotEveryNEvents }}}
-    * Delete Snapshots:
-    *   {{{ (toSequenceNr - 1) - (keepNSnapshots * snapshotEveryNEvents) }}}
-    *
-    * @param toSequenceNr the sequence number to delete to if `deleteEventsOnSnapshot` is false
-    */
-  def toSequenceNumber(lastSequenceNr: Long): Long = {
-    // Delete old events, retain the latest
-    lastSequenceNr - (keepNSnapshots * snapshotEveryNEvents)
-  }
-}
-
-object RetentionCriteria {
-
-  def apply(): RetentionCriteria =
-    RetentionCriteria(
-      snapshotEveryNEvents = 1000L,
-      keepNSnapshots = 2L,
-      deleteEventsOnSnapshot = false)
-
-  /** Scala API. */
-  def apply(snapshotEveryNEvents: Long, keepNSnapshots: Long): RetentionCriteria =
-    RetentionCriteria(snapshotEveryNEvents, keepNSnapshots, deleteEventsOnSnapshot = false)
-
-  /** Java API. */
-  def create(snapshotEveryNEvents: Long, keepNSnapshots: Long): RetentionCriteria =
-    apply(snapshotEveryNEvents, keepNSnapshots)
-
-  /** Java API. */
-  def create(snapshotEveryNEvents: Long, keepNSnapshots: Long, deleteMessagesOnSnapshot: Boolean): RetentionCriteria =
-    apply(snapshotEveryNEvents, keepNSnapshots, deleteMessagesOnSnapshot)
 }
